mesh import : 索引个数计算错误的问题，uint16t越界了，溢出了

+++ Engine/Runtime/RenderSystem/include/mesh/MeshAssimpImpoter.cpp
@@ -1092,7 +1092,7 @@ void MeshAssimpImpoter::processIndice()
     for (auto mesh : mSubMeshs)
     {
         // 处理顶点索引
-        uint16_t indexCount = 0;
+        uint32_t indexCount = 0;
         for(unsigned int i = 0; i < mesh->mNumFaces; i++)
         {
             const aiFace& face = mesh->mFaces[i];
